documentation checkpoint
