+++ interp/src/structures/environment.rs
@@ -6,21 +6,19 @@ use super::names::{
 };
 use super::stk_env::StackMap;
 use crate::configuration::Config;
-use crate::debugger::{name_tree::ActiveTreeNode, PrintCode};
+use crate::debugger::name_tree::ActiveTreeNode;
 use crate::errors::{InterpreterError, InterpreterResult};
 use crate::interpreter::{ComponentInterpreter, Interpreter};
 use crate::interpreter_ir as iir;
-use crate::primitives::{
-    combinational, stateful, Entry, Primitive, Serializeable,
-};
+use crate::primitives::{combinational, stateful, Primitive};
+use crate::structures::state_views::StateView;
 use crate::{
     utils::{AsRaw, MemoryMap},
     values::Value,
 };
 use calyx::ir::{self, RRC};
-use serde::Serialize;
 use std::cell::RefCell;
-use std::collections::{BTreeMap, HashMap, HashSet};
+use std::collections::{HashMap, HashSet};
 use std::iter::once;
 use std::rc::Rc;
 
@@ -34,8 +32,7 @@ type ConstPort = *const ir::Port;
 
 /// A map defining primitive implementations for Cells. As it is keyed by
 /// ConstCell the lifetime of the keys is independent of the actual cells.
-pub(crate) type PrimitiveMap =
-    RRC<HashMap<ConstCell, Box<dyn crate::primitives::Primitive>>>;
+pub(crate) type PrimitiveMap = RRC<HashMap<ConstCell, Box<dyn Primitive>>>;
 
 /// A map defining values for ports. As it is keyed by ConstPort, the lifetime of
 /// the keys is independent of the ports. However as a result it is flat, rather
@@ -615,8 +612,9 @@ impl InterpreterState {
     }
 
     pub fn as_state_view(&self) -> StateView<'_> {
-        StateView::SingleView(self)
+        self.into()
     }
+
     pub fn get_active_tree(&self) -> Vec<ActiveTreeNode> {
         let lookup = self.cell_map.borrow();
 
@@ -626,298 +624,5 @@ impl InterpreterState {
                 lookup[x].get_comp_interpreter().unwrap().get_active_tree()
             })
             .collect()
-    }
-}
-
-impl Serialize for InterpreterState {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        let sv: StateView = self.into();
-        sv.gen_serialzer(false).serialize(serializer)
-    }
-}
-#[allow(clippy::borrowed_box)]
-#[derive(Serialize, Clone)]
-/// Struct to fully serialize the internal state of the environment
-pub struct FullySerialize {
-    ports: BTreeMap<ir::Id, BTreeMap<ir::Id, BTreeMap<ir::Id, Entry>>>,
-    memories: BTreeMap<ir::Id, BTreeMap<ir::Id, Serializeable>>,
-}
-
-#[derive(Clone)]
-pub struct CompositeView<'a>(&'a InterpreterState, Vec<StateView<'a>>);
-
-impl<'a, 'outer> CompositeView<'a> {
-    pub fn new(state: &'a InterpreterState, vec: Vec<StateView<'a>>) -> Self {
-        Self(state, vec)
-    }
-}
-
-impl<'a> Serialize for StateView<'a> {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        self.gen_serialzer(false).serialize(serializer)
-    }
-}
-
-#[derive(Clone)]
-pub enum StateView<'inner> {
-    SingleView(&'inner InterpreterState),
-    Composite(CompositeView<'inner>),
-}
-
-impl<'a, 'outer> From<&'a InterpreterState> for StateView<'a> {
-    fn from(env: &'a InterpreterState) -> Self {
-        Self::SingleView(env)
-    }
-}
-
-impl<'a> From<CompositeView<'a>> for StateView<'a> {
-    fn from(cv: CompositeView<'a>) -> Self {
-        Self::Composite(cv)
-    }
-}
-
-impl<'a> StateView<'a> {
-    pub fn lookup<P: AsRaw<ir::Port>>(&self, target: P) -> &Value {
-        match self {
-            StateView::SingleView(sv) => sv.get_from_port(target),
-            StateView::Composite(cv) => match cv.1.len() {
-                0 => cv.0.get_from_port(target),
-                1 => cv.1[0].lookup(target),
-                _ => {
-                    let original = cv.0.get_from_port(target.as_raw());
-                    let new =
-                        cv.1.iter()
-                            .filter_map(|x| {
-                                let val = x.lookup(target.as_raw());
-                                if val == original {
-                                    None
-                                } else {
-                                    Some(val)
-                                }
-                            })
-                            .collect::<Vec<_>>();
-                    match new.len() {
-                        0 => original,
-                        1 => new[0],
-                        _ => panic!("conflicting parallel values"),
-                    }
-                }
-            },
-        }
-    }
-
-    pub fn sub_component_currently_executing(&self) -> HashSet<GroupQIN> {
-        match self {
-            StateView::SingleView(sv) => sv.sub_component_currently_executing(),
-            StateView::Composite(c) => c.0.sub_component_currently_executing(),
-        }
-    }
-
-    pub fn get_ctx(&self) -> &iir::ComponentCtx {
-        match self {
-            StateView::SingleView(sv) => &sv.context,
-            StateView::Composite(cv) => &cv.0.context,
-        }
-    }
-
-    pub fn get_cell_map(&self) -> &PrimitiveMap {
-        match self {
-            StateView::SingleView(sv) => &sv.cell_map,
-            StateView::Composite(cv) => &cv.0.cell_map,
-        }
-    }
-
-    pub fn get_comp(&self) -> &Rc<iir::Component> {
-        match self {
-            StateView::SingleView(c) => &c.component,
-            StateView::Composite(c) => &c.0.component,
-        }
-    }
-    pub fn get_active_tree(&self) -> Vec<ActiveTreeNode> {
-        match self {
-            StateView::SingleView(c) => c.get_active_tree(),
-            StateView::Composite(c) => c.0.get_active_tree(),
-        }
-    }
-
-    pub fn get_cell_state<R: AsRaw<ir::Cell>>(
-        &self,
-        cell: R,
-        print_code: &PrintCode,
-    ) -> Serializeable {
-        let map = self.get_cell_map();
-        let map_ref = map.borrow();
-        map_ref
-            .get(&cell.as_raw())
-            .map(|x| Primitive::serialize(&**x, Some(*print_code)))
-            .unwrap_or(Serializeable::Empty)
-    }
-
-    /// Returns a string representing the current state of the environment. This
-    /// just serializes the environment to a string and returns that string
-    pub fn state_as_str(&self) -> String {
-        serde_json::to_string_pretty(&self.gen_serialzer(false)).unwrap()
-    }
-
-    pub fn get_cells<S: AsRef<str> + Clone>(
-        &self,
-        name: &S,
-    ) -> Vec<RRC<ir::Cell>> {
-        let ctx_ref = self.get_ctx();
-        ctx_ref.iter().filter_map(|x| x.find_cell(name)).collect()
-    }
-
-    pub fn get_cell<S: AsRef<str> + Clone>(
-        &self,
-        name: S,
-    ) -> Option<RRC<ir::Cell>> {
-        match self {
-            StateView::SingleView(sv) => sv.component.find_cell(&name),
-            StateView::Composite(cv) => cv.0.component.find_cell(&name),
-        }
-    }
-
-    pub fn gen_serialzer(&self, raw: bool) -> FullySerialize {
-        let ctx = self.get_ctx();
-        let cell_prim_map = &self.get_cell_map().borrow();
-
-        let bmap: BTreeMap<_, _> = ctx
-            .iter()
-            .filter(|x| x.name == self.get_comp().name) // there should only be one such comp
-            .map(|comp| {
-                let inner_map: BTreeMap<_, _> = comp
-                    .cells
-                    .iter()
-                    .map(|cell| {
-                        let inner_map: BTreeMap<_, _> = cell
-                            .borrow()
-                            .ports
-                            .iter()
-                            .map(|port| {
-                                let value = self.lookup(port.as_raw());
-
-                                (
-                                    port.borrow().name.clone(),
-                                    if port
-                                        .borrow()
-                                        .attributes
-                                        .has("interp_signed")
-                                    {
-                                        value.as_i64().into()
-                                    } else {
-                                        value.as_u64().into()
-                                    },
-                                )
-                            })
-                            .collect();
-                        (cell.borrow().name().clone(), inner_map)
-                    })
-                    .collect();
-                (comp.name.clone(), inner_map)
-            })
-            .collect();
-        let cell_map: BTreeMap<_, _> = ctx
-            .iter()
-            .filter(|x| x.name == self.get_comp().name)
-            .map(|comp| {
-                let inner_map: BTreeMap<_, _> = comp
-                    .cells
-                    .iter()
-                    .filter_map(|cell_ref| {
-                        let cell = cell_ref.borrow();
-                        if cell.get_attribute("external").is_some() {
-                            if let Some(prim) = cell_prim_map
-                                .get(&(&cell as &ir::Cell as ConstCell))
-                            {
-                                if !prim.is_comb() {
-                                    return Some((
-                                        cell.name().clone(),
-                                        Primitive::serialize(
-                                            &**prim,
-                                            raw.then(|| PrintCode::Binary),
-                                        ), //TODO Griffin: Fix this
-                                    ));
-                                }
-                            }
-                        }
-                        None
-                    })
-                    .collect();
-                (comp.name.clone(), inner_map)
-            })
-            .collect();
-
-        FullySerialize {
-            ports: bmap,
-            memories: cell_map,
-        }
-    }
-}
-
-pub struct MutCompositeView<'a>(
-    &'a mut InterpreterState,
-    Vec<MutStateView<'a>>,
-);
-
-pub enum MutStateView<'inner> {
-    Single(&'inner mut InterpreterState),
-    Composite(MutCompositeView<'inner>),
-}
-
-impl<'inner> MutCompositeView<'inner> {
-    pub fn new(
-        state: &'inner mut InterpreterState,
-        vec: Vec<MutStateView<'inner>>,
-    ) -> Self {
-        Self(state, vec)
-    }
-    pub fn insert<P: AsRaw<ir::Port>>(&mut self, port: P, value: Value) {
-        let raw = port.as_raw();
-        self.0.insert(raw, value.clone());
-        for view in self.1.iter_mut() {
-            view.insert(raw, value.clone())
-        }
-    }
-}
-
-impl<'a> From<&'a mut InterpreterState> for MutStateView<'a> {
-    fn from(env: &'a mut InterpreterState) -> Self {
-        Self::Single(env)
-    }
-}
-
-impl<'a> From<MutCompositeView<'a>> for MutStateView<'a> {
-    fn from(mv: MutCompositeView<'a>) -> Self {
-        Self::Composite(mv)
-    }
-}
-
-impl<'a> MutStateView<'a> {
-    pub fn insert<P: AsRaw<ir::Port>>(&mut self, port: P, value: Value) {
-        match self {
-            MutStateView::Single(s) => s.insert(port, value),
-            MutStateView::Composite(c) => c.insert(port, value),
-        }
-    }
-}
-
-pub trait State {
-    fn lookup(&self, target: &*const ir::Port) -> &Value;
-    fn state_as_str(&self) -> String;
-}
-
-impl<'a> State for StateView<'a> {
-    fn lookup(&self, target: &*const ir::Port) -> &Value {
-        StateView::lookup(self, *target)
-    }
-
-    fn state_as_str(&self) -> String {
-        StateView::state_as_str(self)
     }
 }
