+++ interp/src/structures/mod.rs
@@ -1,4 +1,5 @@
 pub mod environment;
 pub mod names;
+pub mod state_views;
 pub mod stk_env;
 pub mod values;
